Move query_agreement_events to Market
